+++ eng/config/PublishData.json
@@ -60,6 +60,7 @@
     "Microsoft.CodeAnalysis.ExternalAccess.AspNetCore": "vs-impl",
     "Microsoft.CodeAnalysis.ExternalAccess.RazorCompiler": "vs-impl",
     "Microsoft.CodeAnalysis.ExternalAccess.VisualDiagnostics": "vs-impl",
+    "Microsoft.CodeAnalysis.ExternalAccess.TestDiscovery": "vs-impl",
     "Microsoft.CodeAnalysis.ExternalAccess.Xaml": "vs-impl",
     "Microsoft.CodeAnalysis.LanguageServer.ExternalAccess.Copilot": "vs-impl",
     "Microsoft.CodeAnalysis.Remote.ServiceHub": "vs-impl",
